Separated tests into different files
